fmp/requirements.txt
-i https://mirrors.aliyun.com/pypi/simple/

Django==3.1.7
numpy==1.20.1
pandas== 1.2.2
matplotlib==3.3.4
scikit-Learn==0.24.1
xgboost==1.3.3
seaborn==0.11.1
joblib==1.0.1
djangorestframework==3.11.0
django-simpleui==2021.3
django-import-export==2.5.0
